Modified bluecove to work with OS X 10.8
The main part of the code was adapted to use the new bluetooth APIs introduced by Apple and which were enforced with the release of OS X 10.8 (aka Mountain Lion). The modified code facilitates GCD and semaphores to overcome the (also) deprecated concurrency APIs. However, still some function calls have no counterpart in the new API and were disabled for the moment.